move payload parser to usecases
